chore: move common components and hooks
